feat(modules): implement cjsToEsm task
